Add ci mode to evals cli
